% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectBDData.R
\name{CollectBDData}
\alias{CollectBDData}
\title{Consolidate 10K data into one table}
\usage{
CollectBDData()
}
\value{
Writes a final data table with all downloaded 10K information to the location specified in the code
}
\description{
Consolidate 10K data into one table
}
\examples{
CollectBDData()
}
