% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCompanyList.R
\name{GetCompanyList}
\alias{GetCompanyList}
\title{Get the latest company list for NASDAQ####}
\usage{
GetCompanyList(Exchange = "NASDAQ")
}
\arguments{
\item{Exchange}{Defaults to "NASDAQ". Other possible values are "NYSE" & "AMEX"}
}
\value{
A dataframe with the company list associated with the exchange requested
}
\description{
Get the latest company list for NASDAQ####
}
\examples{
GetCompanyList(Exchange = "NASDAQ")
}
