% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTenKs.R
\name{GetTenKs}
\alias{GetTenKs}
\title{Get 10K data for specified CIKs for years 2008 - 2018}
\usage{
GetTenKs(CIK_Index = 1:length(CIKs))
}
\arguments{
\item{CIK_Index}{An index for the number of different CIKs to pull}
}
\value{
Will output text files to location specified in the newEDGAR() function
}
\description{
Get 10K data for specified CIKs for years 2008 - 2018
}
\examples{
GetTenKs(CIK_Index = 1:100)
}
