% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newEDGAR.R
\name{newEDGAR}
\alias{newEDGAR}
\title{Modified version of edgar::getBusinDescr() function}
\usage{
newEDGAR(cik.no, filing.year)
}
\arguments{
\item{cik.no}{The Central Index Key given as a number}

\item{filing.year}{A numeric year or vector of numeric years}
}
\value{
The business description for the given CIKs and years
}
\description{
Modified version of edgar::getBusinDescr() function
}
\details{
This function is originally from the edgar package...it has been adjusted to deal with the error: "Error in writeLines(product.descr2, filename2) : object 'product.descr2' not found "
}
\examples{
newEDGAR(cik = 6201, filing.year = c(2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018))
}
